IventoryPlus/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==43.0.3
Django==5.0.7
django-multiselectfield==0.1.13
django-phone-field==1.8.1
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
openpyxl==3.1.5
packaging==24.1
pillow==10.4.0
plotly==5.23.0
psycopg2==2.9.9
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
python-dotenv==1.0.1
sqlparse==0.5.1
tenacity==8.5.0
